Refactor: 차트 api & 생성 파일 분리
