--- src/api/upbit/foo.ts
@@ -1,1 +1,0 @@
-export {};
