+++ src/api/upbit/useWsTicker.ts
@@ -1,0 +1,14 @@
+// import { ITicker, ImarketCodes, TKOptionsInterface } from '../interfaces';
+// import { useRef, useState, useEffect } from 'react';
+// import getLastBuffers from '../functions/getLastBuffers';
+// import sortBuffers from '../functions/sortBuffers';
+// import { throttle } from 'lodash';
+// import socketDataEncoder from '../functions/socketDataEncoder';
+// import updateSocketData from '../functions/updateSocketData';
+// import isArrayOfImarketCodes from '../functions/isArrayOfImarketCodes';
+
+function useWsTicker() {
+  return null;
+}
+
+export default useWsTicker;
