+++ src/pages/home/Home.page.tsx
@@ -3,10 +3,16 @@ import { useRecoilState } from 'recoil';
 import { Table } from 'components/Table';
 import { ChartRight } from 'components/ChartRight';
 import { ChartLeft } from 'components/ChartLeft';
-import { useFetchMarketCode } from 'use-upbit-api';
 import { marketCodesState } from 'recoil/atoms/upbit';
 import { DisplayBoard } from './Home.styles';
 import styled from 'styled-components';
+import useFetchMarketCode from 'api/upbit/useFetchMarketCode';
+
+interface FetchedMCData {
+  market: string;
+  korean_name: string;
+  english_name: string;
+}
 
 const ChartsWrapper = styled.div`
   background-color: whitesmoke;
@@ -21,7 +27,7 @@ export const Home: React.FC = () => {
   const [marketCodes, setMarketCodes] = useRecoilState(marketCodesState);
 
   useEffect(() => {
-    const MarketCodes_KRW = fetchedMC.filter((code) =>
+    const MarketCodes_KRW = fetchedMC.filter((code: FetchedMCData) =>
       code.market.includes('KRW'),
     );
     setMarketCodes(MarketCodes_KRW);
